requirements.txt
asgiref==3.3.4
autopep8==1.5.7
boto3==1.34.71
botocore==1.34.71
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.3
django-admin-inline-paginator==0.4.0
django-appconf==1.0.4
django-braces==1.14.0
django-cleanup==8.1.0
django-cors-headers==3.7.0
django-cryptography==1.0
django-oauth-toolkit==1.5.0
django-six==1.0.4
django-social-share==2.2.1
django-storages==1.14.2
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
ecdsa==0.17.0
idna==2.10
install==1.3.4
jmespath==1.0.1
jwcrypto==0.9
oauthlib==3.1.1
pillow==10.2.0
psycopg2==2.9.1
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.9.0.post0
python-decouple==3.4
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.10.1
setuptools==69.1.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
twilio==7.1.0
urllib3==1.26.5
wheel==0.42.0
wrapt==1.16.0
